requirements.txt
google-adk>=1.0.0
a2a-sdk>=0.3.0
uvicorn[standard]
httpx
python-dotenv
mcp
tavily-python
